python_training/requirements.txt
tensorflow-gpu
pillow
scikit-image
scipy